Rearranged files into (probably) their final spot
