+++ src/components/builder-mode/SelectionMenu.jsx
@@ -13,15 +13,15 @@ import { IoWarningOutline } from "react-icons/io5";
 import { LuSwords } from "react-icons/lu";
 import { RxCross1 } from "react-icons/rx";
 import { v4 as uuid } from "uuid";
-import hero_constraint_data from "../assets/data/hero_constraint_data.json";
-import mesbg_data from "../assets/data/mesbg_data.json";
-import { useStore } from "../state/store";
-import { SelectionSiege } from "./SelectionSiege";
-import { SelectionUnit } from "./SelectionUnit";
-import { allianceColours } from "./constants/alliances";
-import { wanderers } from "./constants/wanderers";
-import { FactionLogo } from "./images/FactionLogo.tsx";
-import { SidebarTypes } from "./sidebar-drawer/sidebars";
+import hero_constraint_data from "../../assets/data/hero_constraint_data.json";
+import mesbg_data from "../../assets/data/mesbg_data.json";
+import { useStore } from "../../state/store";
+import { allianceColours } from "../constants/alliances";
+import { wanderers } from "../constants/wanderers";
+import { FactionLogo } from "../images/FactionLogo.tsx";
+import { SidebarTypes } from "../sidebar-drawer/sidebars";
+import { SelectionSiege } from "./selection/SelectionSiege.jsx";
+import { SelectionUnit } from "./selection/SelectionUnit.jsx";
 
 /* The menu component on the left-hand side used for displaying information about warnings,
 bow limits, and army bonuses. Also used as the selection menu when choosing a unit. */
@@ -42,7 +42,6 @@ export function SelectionMenu({
   factionBowCounts,
   factionModelCounts,
   factionData,
-  setShowKeywordSearch,
 }) {
   const {
     roster,
@@ -205,7 +204,7 @@ export function SelectionMenu({
             <Button
               variant="light"
               className="ms-auto border shadow-sm"
-              onClick={() => setShowKeywordSearch(true)}
+              onClick={() => openSidebar(SidebarTypes.KEYWORD_SEARCH)}
             >
               <FaSearch /> Search Keywords
             </Button>
