+++ src/components/builder-mode/selection/SelectionSiege.jsx
@@ -2,9 +2,9 @@ import Accordion from "react-bootstrap/Accordion";
 import Button from "react-bootstrap/Button";
 import Stack from "react-bootstrap/Stack";
 import { v4 as uuid } from "uuid";
-import siege_equipment from "../assets/data/siege_equipment.json";
-import { useStore } from "../state/store";
-import { UnitProfilePicture } from "./images/UnitProfilePicture.tsx";
+import siege_equipment from "../../../assets/data/siege_equipment.json";
+import { useStore } from "../../../state/store";
+import { UnitProfilePicture } from "../../images/UnitProfilePicture.tsx";
 
 export function SelectionSiege({
   newWarriorFocus,
